Create parseTCGABarcode.R
This is a function that will parse out the tumor/normal tissue information from a list of TCGA barcodes.  It was part of another script, but I'm posting it in case it is useful for others.  In the future it could be enhanced by parsing out other information as well.