updated scripts to fix bug and work with GDAC data
Major update of the script to correct a bug:  I was not converting the
factors to numeric correctly, so it was changing my entire result!

I also added in changes to make it compatible with the new GDAC source
of information.  The survival data now needs to be contained in the
expression data object.

I also added the option to use incremental steps or a set number of
steps for the cutoff selection.